+++ src/main/java/domain/card/CardSelector.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public interface CardSelector {
