+++ src/test/java/domain/card/DeckTest.java
@@ -1,6 +1,5 @@
 package domain.card;
 
-import domain.CardSelector;
 import java.util.HashSet;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
